Add support for STLPort
